Streamline and fix ttyutils code
ttyutils.Winsize is broken (at least on Darwin) and causes a panic.
I'm guessing this has something to do with a change in the memory
layout of structs since it was written. The syscall wants a `[4]uint16`
but the code was passing a struct with two `uint16` members. Maybe when
this was originally written the struct was padded in memory to occupy
the correct 64 bits? I based the new implementation on:
http://godoc.org/src/golang.org/x/crypto/ssh/terminal/util.go

In the process I noticed that essentially all of the code is shared
between Linux and Darwin -- just the constants differ. The x/crypto/ssh/terminal
package provides a good example for how to resolve this and I've followed
that pattern to remove all the duplicated code.

I added tests for a couple methods to make sure I ended up with something
sane and to provide a template for writing regression tests for future bugs.
I wasn't quite sure how to test a terminal package without a terminal
but the `pty` package I had lying around in my GOPATH seemed to work
well. Alternative suggestions welcome!